+++ src/js/insight-globals.ts
@@ -1,4 +1,3 @@
 import { InsightModules } from '../types';
 export const VDL = global.VDL;
-
 export const insightModules: InsightModules = global.insightModules;