+++ src/types/index.ts
@@ -1,21 +1,28 @@
-export interface SetNameAndPosition {
+export type SetNameAndPosition = {
     name: string;
     position: number;
-}
+};
 
-export interface FilterPositionAndValue {
+export type FilterPositionAndValue = {
     index: number;
     value: any;
-}
+};
 
-export interface DataUtils {
+export type IndexFilter = {
+    setName: string;
+    setPosition: number;
+    value: number | string | Array<number | string>;
+};
+
+export type DataUtils = {
     getFilterPositionsAndValues(filters: any, indexSetsAndPositions: SetNameAndPosition[]): FilterPositionAndValue[];
     getSetNamesAndPosns(indexSets: string[]): SetNameAndPosition[];
     entityTypeIsNumber(entity: any): boolean;
-}
+};
 
-export interface InsightModules {
+export type InsightModules = {
     load(mod: 'external-libs/knockout'): KnockoutStatic;
+    load(mod: 'external-libs/jquery'): JQueryStatic;
     load(mod: 'utils/data-utils'): DataUtils;
     load(mod: 'vdl-validator-registry'): any;
     load(mod: 'vdl/vdl-validator-factory'): any;
@@ -27,4 +34,4 @@ export interface InsightModules {
     load(mod: 'components/autotable-select-options'): any;
     load(mod: 'data/set-sorter'): any;
     load(mod: string): any;
-}
+};